+++ frontend/src/components/common/smart/employee/employee.scss
