+++ frontend/src/components/common/smart/leaves/leaves.tsx
@@ -1,18 +1,11 @@
 import { useState, useEffect } from "react";
-import {
-  Box,
-  Typography,
-  InputAdornment,
-  TextField,
-  Button,
-  Tooltip,
-} from "@mui/material";
+import { Box, Typography, Button, Tooltip } from "@mui/material";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
 import moment from "moment";
 import FilterIcon from "@src/assets/svgs/filterButtonIcon.svg";
 import NotfoundIcon from "@src/assets/svgs/requestIcon.svg";
-import searchBox from "@src/assets/svgs/searchBox.svg";
 import LeavesCountBar from "@src/components/common/presentational/leavesCountBar/leavesCountBar";
+import Input from "@src/components/shared/formControls/textInput/textInput";
 import RowSkeletonCard from "@src/components/shared/loaders/rowSkeletonCard/RowSkeletonCard";
 import LeaveModal from "@src/components/shared/popUps/leaveTypeModal/leaveTypeModal";
 import { employeeConstants } from "@src/helpers/constants/constants";
@@ -259,9 +252,6 @@ function Leaves() {
       setLeavesData(rows);
     }
   }, [debouncedSearchTerm, rows]);
-  const handleInput = (e: { target: { value: string } }) => {
-    setQuery(e.target.value);
-  };
   return (
     <Box className="leavesDataGridTable-section">
       <Box
@@ -273,46 +263,7 @@ function Leaves() {
           className="filter-section"
           sx={{ display: "flex", justifyContent: "space-between" }}
         >
-          <Box className="text-cls">
-            <TextField
-              className="searchbox"
-              id="search-headbox"
-              variant="outlined"
-              onChange={handleInput}
-              placeholder="Search Employee here"
-              sx={{
-                "& label.Mui-focused": {
-                  color: "#999999",
-                },
-                "& .MuiInput-underline:after": {
-                  borderBottomColor: "#E7E7E7",
-                },
-                "& .MuiOutlinedInput-root": {
-                  "& fieldset": {
-                    borderColor: "#E7E7E7",
-                  },
-                  "&:hover fieldset": {
-                    borderColor: "#999999",
-                  },
-                  "&.Mui-focused fieldset": {
-                    borderColor: "#999999",
-                  },
-                },
-              }}
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">
-                    {/* SearchBoxSVG */}
-                    <img
-                      className="profile-pic"
-                      src={searchBox}
-                      alt="profile pic"
-                    />
-                  </InputAdornment>
-                ),
-              }}
-            />
-          </Box>
+          <Input setSearchText={setQuery} />
           <Box className="filter-btn-cls">
             <Button
               className="filter-btn"
