+++ frontend/src/components/common/smart/teams/teams.tsx
@@ -1,17 +1,11 @@
 import { useState, useEffect } from "react";
 import AddIcon from "@mui/icons-material/Add";
-import {
-  Box,
-  Typography,
-  InputAdornment,
-  TextField,
-  Button,
-} from "@mui/material";
+import { Box, Typography, Button } from "@mui/material";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
 import moment from "moment";
 import FilterIcon from "@src/assets/svgs/filterButtonIcon.svg";
 import NotfoundIcon from "@src/assets/svgs/requestIcon.svg";
-import searchBox from "@src/assets/svgs/searchBox.svg";
+import Input from "@src/components/shared/formControls/textInput/textInput";
 import RowSkeletonCard from "@src/components/shared/loaders/rowSkeletonCard/RowSkeletonCard";
 import CreateTeamModal from "@src/components/shared/popUps/createTeam/createTeam";
 import { employeeConstants } from "@src/helpers/constants/constants";
@@ -34,7 +28,7 @@ function Teams() {
   const [query, setQuery] = useState("");
   const { notFound } = constantData.Employee;
   const debouncedSearchTerm = useDebounce(query, 500);
-  const { teams, AddTeam } = constantData.Teams;
+  const { Teams, AddTeam } = constantData.Teams;
   const { filterButton } = constantData.Buttons;
   const [teamsData, setTeamsData] = useState<teamTypes[]>([]);
   const [pageSize, setPageSize] = useState<number>(10);
@@ -120,9 +114,6 @@ function Teams() {
       setTeamsData(rows);
     }
   }, [debouncedSearchTerm, rows]);
-  const handleInput = (e: { target: { value: string } }) => {
-    setQuery(e.target.value);
-  };
 
   return (
     <Box className="teamsDataGridTable-section">
@@ -130,51 +121,12 @@ function Teams() {
         className="top-bar-cls"
         sx={{ display: "flex", justifyContent: "space-between" }}
       >
-        <Typography className="title-cls">{teams}</Typography>
+        <Typography className="title-cls">{Teams}</Typography>
         <Box
           className="filter-section"
           sx={{ display: "flex", justifyContent: "space-between" }}
         >
-          <Box className="text-cls">
-            <TextField
-              className="searchbox"
-              id="search-headbox"
-              variant="outlined"
-              onChange={handleInput}
-              placeholder="Search Employee here"
-              sx={{
-                "& label.Mui-focused": {
-                  color: "#999999",
-                },
-                "& .MuiInput-underline:after": {
-                  borderBottomColor: "#E7E7E7",
-                },
-                "& .MuiOutlinedInput-root": {
-                  "& fieldset": {
-                    borderColor: "#E7E7E7",
-                  },
-                  "&:hover fieldset": {
-                    borderColor: "#999999",
-                  },
-                  "&.Mui-focused fieldset": {
-                    borderColor: "#999999",
-                  },
-                },
-              }}
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">
-                    {/* SearchBoxSVG */}
-                    <img
-                      className="profile-pic"
-                      src={searchBox}
-                      alt="profile pic"
-                    />
-                  </InputAdornment>
-                ),
-              }}
-            />
-          </Box>
+          <Input setSearchText={setQuery} />
           <Box className="filter-btn-cls">
             <Button
               className="filter-btn"
