+++ frontend/src/components/shared/formControls/textInput/textInput.scss
@@ -1,0 +1,13 @@
+.text-cls {
+  margin-right: 15px;
+  border: 1px solid white;
+  .MuiInputBase-root {
+    padding: 11px 16px 11px 16px;
+    border-radius: 12px;
+    height: 44px;
+    background: #ffffff;
+    box-shadow: 10px 10px 24px 1px rgba(221, 221, 221, 0.15);
+    border-radius: 12px;
+    border: 1px solid #ffffff;
+  }
+}
