+++ frontend/src/components/shared/popUps/createTeam/createTeam.tsx
@@ -1,4 +1,4 @@
-// import { useState } from "react";
+import { useState } from "react";
 import CheckBoxIcon from "@mui/icons-material/CheckBox";
 import CheckBoxOutlineBlankIcon from "@mui/icons-material/CheckBoxOutlineBlank";
 import { LoadingButton } from "@mui/lab";
@@ -12,17 +12,20 @@ import DialogContent from "@mui/material/DialogContent";
 import DialogTitle from "@mui/material/DialogTitle";
 import TextField from "@mui/material/TextField";
 import { useFormik } from "formik";
-// import { toast } from "react-toastify";
+import { toast } from "react-toastify";
 import * as yup from "yup";
 import crossIcon from "@src/assets/svgs/cross.svg";
 import submitIcon from "@src/assets/svgs/Frame.svg";
-// import { timeOut } from "@src/helpers/constants/constants";
+import { timeOut } from "@src/helpers/constants/constants";
 import "@src/components/shared/popUps/createTeam/createTeam.scss";
 import { LocalizationInterface } from "@src/helpers/interfaces/localizationinterfaces";
 import { localizedData } from "@src/helpers/utils/language";
-// import { toastAPIError } from "@src/helpers/utils/utils";
+import { toastAPIError } from "@src/helpers/utils/utils";
 
-import { useGetEmployeesQuery } from "@src/store/reducers/employees-api";
+import {
+  useGetEmployeesQuery,
+  useCreateTeamMutation,
+} from "@src/store/reducers/employees-api";
 
 interface Props {
   open: boolean;
@@ -33,8 +36,9 @@ const checkedIcon = <CheckBoxIcon fontSize="small" />;
 
 const CreateTeamModal = ({ open, handleClose }: Props) => {
   const constantData: LocalizationInterface = localizedData();
-  // const [loading, setLoading] = useState(false);
+  const [loading, setLoading] = useState(false);
   const { data: rows = [] } = useGetEmployeesQuery();
+  const [createTeam] = useCreateTeamMutation();
   const { cancelBtn, saveBtn } = constantData.Modals;
   const {
     CreateTeam,
@@ -42,51 +46,54 @@ const CreateTeamModal = ({ open, handleClose }: Props) => {
     NameRequired,
     Name,
     AddTeamMember,
-    // MemberRequired,
+    MemberRequired,
   } = constantData.Teams;
 
   const formik = useFormik({
     initialValues: {
       name: "",
+      members: [],
     },
     validationSchema: yup.object({
       name: yup.string().required(NameRequired),
+      members: yup.array().min(1, MemberRequired),
     }),
     validateOnChange: true,
     onSubmit: () => {
-      // handleCreateStandup();
+      handleCreateTeam();
     },
   });
   const resetForm = () => {
     handleClose();
     formik.resetForm();
   };
-  // const handleCreateStandup = async () => {
-  //   setLoading(true);
-  //   const standupObj = getCreateStandupObject();
-  //   await createStandup({ standupObject: standupObj })
-  //     .unwrap()
-  //     .then(async () => {
-  //       toast.success("Standup successfully added.", {
-  //         autoClose: timeOut,
-  //         pauseOnHover: false,
-  //       });
-  //       formik.resetForm();
-  //       handleClose();
-  //       setLoading(false);
-  //     })
-  //     .catch((error) => {
-  //       setLoading(false);
-  //       toastAPIError("Something went wrong.", error.status, error.data);
-  //     });
-  // };
-  // const getCreateStandupObject = () => {
-  //   return {
-  //     team: formik.values.team,
-  //     name: formik.values.name,
-  //     time: moment(formik.values.time).format("HH:mm:ss"),
-  //   };
-  // };
+
+  const handleCreateTeam = async () => {
+    setLoading(true);
+    const teamObj = getCreateTeamObject();
+    await createTeam({ teamObject: teamObj })
+      .unwrap()
+      .then(async () => {
+        toast.success("Team successfully added.", {
+          autoClose: timeOut,
+          pauseOnHover: false,
+        });
+        formik.resetForm();
+        handleClose();
+        setLoading(false);
+      })
+      .catch((error) => {
+        setLoading(false);
+        toastAPIError("Something went wrong.", error.status, error.data);
+      });
+  };
+  const getCreateTeamObject = () => {
+    const extractedIDS = formik.values.members.map((obj) => obj?.id);
+    return {
+      name: formik.values.name,
+      members: extractedIDS,
+    };
+  };
   return (
     <>
       <Dialog open={open} onClose={resetForm} className="createTeamModal">
@@ -125,12 +132,15 @@ const CreateTeamModal = ({ open, handleClose }: Props) => {
               multiple
               options={rows}
               className="text-field-cls"
+              onChange={(e, value) =>
+                formik.setFieldValue("members", value || [])
+              }
               disableCloseOnSelect
               getOptionLabel={(option) =>
                 `${option?.first_name} ${option?.last_name}`
               }
               renderOption={(props, option, { selected }) => (
-                <li {...props}>
+                <li {...props} key={option?.id}>
                   <Checkbox
                     icon={icon}
                     checkedIcon={checkedIcon}
@@ -143,13 +153,14 @@ const CreateTeamModal = ({ open, handleClose }: Props) => {
               renderInput={(params) => (
                 <TextField
                   {...params}
+                  required
                   label={AddTeamMember}
                   InputLabelProps={{ className: "textfield_label" }}
                 />
               )}
             />
             <Typography className="errorText">
-              {formik.touched.name && formik.errors.name}
+              {formik.touched.members && formik.errors.members}
             </Typography>
           </Box>
         </DialogContent>
