+++ frontend/src/store/reducers/employees-api.tsx
@@ -32,7 +32,14 @@ export const employeesApi = createApi({
       return headers;
     },
   }),
-  tagTypes: ["Employee", "Owner", "Leaves", "Standup", "CreateStandup"],
+  tagTypes: [
+    "Employee",
+    "Owner",
+    "Leaves",
+    "Standup",
+    "CreateStandup",
+    "Teams",
+  ],
   endpoints: (builder) => ({
     getEmployees: builder.query<Employee[], void>({
       query: () => "/employees/",
@@ -152,6 +159,7 @@ export const employeesApi = createApi({
     }),
     getTeams: builder.query<teamTypes[], void>({
       query: () => "/team/",
+      providesTags: ["Teams"],
     }),
     createStandup: builder.mutation({
       query: ({ standupObject }) => {
@@ -186,6 +194,16 @@ export const employeesApi = createApi({
     getAttendacne: builder.query<AttendanceTypes[], void>({
       query: () => "/attendance/",
     }),
+    createTeam: builder.mutation({
+      query: ({ teamObject }) => {
+        return {
+          url: "/team/",
+          method: "POST",
+          body: teamObject,
+        };
+      },
+      invalidatesTags: ["Teams"],
+    }),
   }),
 });
 
@@ -217,4 +235,5 @@ export const {
   useAddStandupMutation,
   useGetAssetsTypesQuery,
   useGetAttendacneQuery,
+  useCreateTeamMutation,
 } = employeesApi;
